Make centerview work with gyro
Default pitch drift code tries to move towards the ideal pitch
every frame, which means it is always fighting the player's input.
When gyro is active we disable the default behavior, and instead
we perform "additive" camera centering by storing the delta
between the ideal pitch and the current pitch when centerview is used,
and injecting that delta over some amount of time.